Move identification to backend directory
