+++ L1Trigger/L1THGCal/plugins/backend/HGCalTriggerClusterIdentificationBDT.cc
@@ -1,7 +1,7 @@
 #include <limits>
 #include "FWCore/ParameterSet/interface/FileInPath.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "L1Trigger/L1THGCal/interface/be_algorithms/HGCalTriggerClusterIdentificationBase.h"
+#include "L1Trigger/L1THGCal/interface/backend/HGCalTriggerClusterIdentificationBase.h"
 #include "L1Trigger/L1THGCal/interface/HGCalTriggerTools.h"
 #include "CommonTools/Utils/interface/TMVAEvaluator.h"
