+++ L1Trigger/L1THGCal/src/backend/HGCalTriggerClusterIdentificationBase.cc
@@ -1,4 +1,4 @@
-#include "L1Trigger/L1THGCal/interface/be_algorithms/HGCalTriggerClusterIdentificationBase.h"
+#include "L1Trigger/L1THGCal/interface/backend/HGCalTriggerClusterIdentificationBase.h"
 
 EDM_REGISTER_PLUGINFACTORY(HGCalTriggerClusterIdentificationFactory,
                            "HGCalTriggerClusterIdentificationFactory");
